organise mongo package
